Learned abbout dragcontrols + split parts of our canvas into their own components
